Refactored out Alert_MaybeCritical
