+++ MoreAlerts/Alert_Custom.cs
@@ -8,13 +8,16 @@ using Verse;
 
 namespace MoreAlerts
 {
-    abstract class Alert_Custom : Alert_Critical
+    abstract class Alert_Custom : Alert_MaybeCritical
     {
 
         protected List<Thing> affectedThings = new List<Thing>();
         protected int lastTick = 0;
 
-        private int lastActiveFrame = -1;
+        public Alert_Custom()
+        {
+            this.defaultPriority = AlertPriority.Medium;
+        }
 
         public override AlertReport GetReport()
         {
@@ -42,68 +45,6 @@ namespace MoreAlerts
         }
 
         protected abstract void GetAffectedThings();
-
-        public override void AlertActiveUpdate()
-        {
-            /*
-            if (this.defaultPriority == AlertPriority.Critical)
-            {
-                (this as Alert_Critical).AlertActiveUpdate();
-            }
-            else
-            {
-                (this as Alert).AlertActiveUpdate();
-            }
-            */
-
-            if (this.defaultPriority == AlertPriority.Critical)
-            {
-                if (this.lastActiveFrame < Time.frameCount - 1)
-                {
-                    Messages.Message("MessageCriticalAlert".Translate(new object[]
-                    {
-                    this.GetLabel()
-                    }), this.GetReport().culprit, MessageSound.SeriousAlert);
-                }
-                this.lastActiveFrame = Time.frameCount;
-            }
-            else
-            {
-                // Nothing.
-            }
-
-        }
-
-        protected override Color BGColor
-        {
-            /*
-            get
-            {
-                if (this.defaultPriority == AlertPriority.Critical)
-                {
-                    return (this as Alert_Critical).BGColor;
-                }
-                else
-                {
-                    return (this as Alert).BGColor;
-                }
-            }
-            */
-
-            get
-            {
-                if (this.defaultPriority == AlertPriority.Critical)
-                {
-                    float num = Pulser.PulseBrightness(0.5f, Pulser.PulseBrightness(0.5f, 0.6f));
-                    return new Color(num, num, num) * Color.red;
-                }
-                else
-                {
-                    return Color.clear;
-                }
-            }
-
-        }
 
     }
 }
