+++ MoreAlerts/MoreAlerts.csproj
@@ -54,6 +54,7 @@
     <Compile Include="Alert_Custom_PrisonerSpawned.cs" />
     <Compile Include="Alert_Custom_Spawned.cs" />
     <Compile Include="Alert_ImmunityDeath.cs" />
+    <Compile Include="Alert_MaybeCritical.cs" />
     <Compile Include="Alert_NotRecruitingPrisoner.cs" />
     <Compile Include="Alert_PawnCold.cs" />
     <Compile Include="Alert_PawnHot.cs" />
